Created fighting sequence
At the moment it just spams the entire fight. Make it so that there is a timer, and show the health of the hero as well as enemy. The enemy also needs to be placed in an array, so that it can be removed from the list of enemies. There still needs to be a level up system