Updated UI and fighting sequence
Added linespaces between options in UI, to make everyting more readable. The fighting sequence now have a (4 second) timer inbetween actions. A vector has been made to store enemies in, and enemies should be deleted when defeated. The whole enemy interaction should be wrapped in a loop to let the player go through all of the enemies.